implementation/h2o-benchmarker/requirements.txt
autoflake==0.6.6
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
brewer2mpl==1.4.1
certifi==2016.2.28
configparser==3.5.0
cycler==0.10.0
DateTime==4.1.1
decorator==4.0.10
docutils==0.12
entrypoints==0.2.2
functools32==3.2.3.post2
future==0.15.2
ggplot==0.10.4
h2o==3.8.3.99999
ipykernel==4.3.1
ipython==5.0.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
jupyterlab==0.2.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.1
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
patsy==0.4.1
pexpect==4.2.0
pickleshare==0.7.2
prompt-toolkit==1.0.3
psutil==4.3.0
ptyprocess==0.5.1
pyflakes==1.2.3
Pygments==2.1.3
pynisher==0.4.2
pyparsing==2.1.5
pysmac==0.9.1
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
pyzmq==15.3.0
qtconsole==4.2.1
readline==6.2.4.1
requests==2.10.0
rpy2==2.8.2
run==0.2
scikit-learn==0.17.1
scipy==0.17.1
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.14
statsmodels==0.6.1
tabulate==0.7.5
terminado==0.6
tornado==4.3
traitlets==4.2.2
wcwidth==0.1.7
widgetsnbextension==1.2.3
zope.interface==4.2.0
